// src/app/home/klienci/models/klient.ts
export interface Klient {
  _id: string;
  typ_prawny: number;
  n_firmy: string;
  pesel: string;
  nazwisko: string;
  imie: string;
  kod: string;
  miejscowosc: string;
  ulica: string;
  kod_k: string;
  miejscowosc_k: string;
  ulica_k: string;
  tel: string;
  tel_kom: string;
  email: string;
  data_wyd: Date;
  Kat: string;
  nip: string;
  regon: string;
  uwagi: string;
}